rename backspace compare to change id
