+++ Python/problems/problem_0168_backspace_compare.py
